feat: Add user information capture functionality
This commit adds the functionality to capture visitor information on the website. It introduces a new component called CaptureVisitor, which collects data such as user agent, browser, operating system, device type, screen resolution, referrer, language, and session ID. The captured data is then sent to the server using an HTTP POST request to the "/api/userInformation" endpoint. This commit also includes the necessary changes to the route and model files to handle the visitor data.

Refactor the Hero component and add a WhatsApp contact button

This commit optimizes the Hero component by making necessary code improvements. It also adds a WhatsApp contact button to the Hero component, allowing users to easily contact the website owner via WhatsApp.

Add lazy loading for images in layout, VisitorData, and Home components

This commit introduces lazy loading for images in the layout, VisitorData, and Home components. Lazy loading improves the performance of the website by loading images only when they are visible on the screen, reducing the initial page load time.

Add URL for pharmanecia and Apticon projects

This commit adds the URLs for the pharmanecia and Apticon projects. These URLs will be used to link to the respective projects on the website.

Remove unnecessary URL

This commit removes an unnecessary URL from the codebase.

Update dependencies

This commit updates the dependencies in the package.json file. It adds the "@next/swc-darwin-arm64" package and updates other packages to their latest versions.