feat: Add API route for fetching site visitor list
This commit adds a new API route in the `siteVisitorList` module to fetch the list of site visitors from the database. The route is implemented in the `route.ts` file and uses the `SiteVisitorModel` to retrieve the visitor list. If there is an error during the fetching process, an appropriate error response is returned.

The purpose of this change is to provide functionality for retrieving the site visitor list, which can be used in the `VisitorList` component.

Refactor the `VisitorList` component

This commit adds a new component called `VisitorList` in the `site-visitor` module. The component fetches the site visitor list using the newly added API route and displays it in a table format. The component handles loading and error states and renders the visitor data in the table rows.

The purpose of this change is to optimize the `Hero` component and add a WhatsApp contact button to improve user experience.

Add new files for the Pharmanecia and Apticon projects

This commit adds new image files for the Pharmanecia and Apticon projects. The images are added to the respective project directories.

The purpose of this change is to include the necessary image assets for the Pharmanecia and Apticon projects.